Bug Fix - Stylish quotes problem
Nice modern computers use fancy left and right quotes but that causes crashes

e.g.

If your event is something along the lines of ...Watch "The Longest Day"  the program may crash.

So we need to fix this by replacing the nicely styled quotes with the more boring basic quotes